requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
boto3==1.9.156
botocore==1.12.156
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.6.1
dateparser==0.7.1
docutils==0.14
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
numpy==1.16.2
peewee==3.9.4
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-binance==0.7.1
python-bittrex==0.3.0
python-dateutil==2.8.0
pytz==2019.1
regex==2019.4.14
requests==2.22.0
s3transfer==0.2.0
service-identity==18.1.0
six==1.12.0
termcolor==1.1.0
Twisted==19.2.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.25.3
zope.interface==4.6.0
